+++ packages/trader/src/AppV2/Components/AccumulatorStats/accumulator-stats-description.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ActionSheet, Text } from '@deriv-com/quill-ui';
 import { Localize } from '@deriv/translations';
 import { UNIFIED_MODE_VIDEO_ID } from 'Modules/Trading/Helpers/video-config';
-import StreamIframe from '../OnboardingGuide/StreamIframe';
+import StreamIframe from '../StreamIframe';
 
 const AccumulatorStatsDescription = ({ onActionSheetClose }: { onActionSheetClose: () => void }) => {
     return (
