+++ packages/trader/src/AppV2/Components/AccumulatorStats/accumulator-stats.scss
@@ -172,7 +172,7 @@
     }
 
     .stats-description {
-        text-align: left;
+        text-align: start;
 
         &__content {
             margin: var(--core-spacing-1200);
