+++ packages/trader/src/AppV2/Components/StreamIframe/__tests__/stream-iframe.spec.tsx
