+++ packages/trader/src/AppV2/Components/StreamIframe/index.ts
