+++ packages/trader/src/AppV2/Components/StreamIframe/stream-iframe.scss
